Rename code_generator folder to code_generation.
